Set ownership and labels to the populator pod
This patch sets the PVC as the owner of the populator pod and the. It
also labels the OvirtVolumePopulator and the OpenstackVolumePopulator
CRs with the VM ID and the migration ID. This is useful when we will
wish to get the logs using must-gather or lookup for the relevant
populator resources.

The populator pod won't be deleted at the end of migration in order to
let us gather the logs.

Once cleanup is triggered by the controller, the CRs and populator pod
will be deleted.

When the controller starts, it will lookup based on the plan for
OpenstackVolumePopulator and OvirtVolumePopulator CRs, updating the
labels with the VM ID and migration ID. That will be handy when
upgrading to newer version while having old resources, letting them to
be cleaned up.

Owner references added to the Populator CRs and the populator pods only
after the VM is created. In order to have additional clean-up support
mostly on remote clusters when the controller won't have the ability to
do any operation after the migration.

+++ pkg/lib-volume-populator/populator-machinery/controller.go
@@ -31,6 +31,7 @@ import (
 	"time"
 
 	"github.com/konveyor/forklift-controller/pkg/apis/forklift/v1beta1"
+	"github.com/konveyor/forklift-controller/pkg/controller/plan"
 	corev1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -586,13 +587,21 @@ func (c *controller) syncPvc(ctx context.Context, key, pvcNamespace, pvcName str
 				// Join the transfer network namespace and name
 				annotations[AnnDefaultNetwork] = fmt.Sprintf("%s/%s", transferNetwork["namespace"], transferNetwork["name"])
 			}
-
+			migration, found, err := unstructured.NestedString(crInstance.Object, "metadata", "labels", "migration")
+			if err != nil {
+				return err
+			}
+			labels := map[string]string{"pvcName": pvc.Name}
+			if found {
+				labels["migration"] = migration
+			}
 			// Make the pod
 			pod = &corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:        podName,
 					Namespace:   populatorNamespace,
 					Annotations: annotations,
+					Labels:      labels,
 				},
 				Spec: makePopulatePodSpec(pvcPrimeName, secretName),
 			}
@@ -686,9 +695,11 @@ func (c *controller) syncPvc(ctx context.Context, key, pvcNamespace, pvcName str
 			if corev1.PodFailed == pod.Status.Phase {
 				c.recorder.Eventf(pvc, corev1.EventTypeWarning, reasonPodFailed, "Populator failed: %s", pod.Status.Message)
 				// Delete failed pods so we can try again
-				err = c.kubeClient.CoreV1().Pods(populatorNamespace).Delete(ctx, pod.Name, metav1.DeleteOptions{})
-				if err != nil {
-					return err
+				if !plan.RestartLimitExceeded(pod) {
+					err = c.kubeClient.CoreV1().Pods(populatorNamespace).Delete(ctx, pod.Name, metav1.DeleteOptions{})
+					if err != nil {
+						return err
+					}
 				}
 			}
 			// We'll get called again later when the pod succeeds
@@ -760,14 +771,6 @@ func (c *controller) syncPvc(ctx context.Context, key, pvcNamespace, pvcName str
 
 	// *** At this point the volume population is done and we're just cleaning up ***
 	c.recorder.Eventf(pvc, corev1.EventTypeNormal, reasonPodFinished, "Populator finished")
-
-	// If the pod still exists, delete it
-	if pod != nil {
-		err = c.kubeClient.CoreV1().Pods(populatorNamespace).Delete(ctx, pod.Name, metav1.DeleteOptions{})
-		if err != nil {
-			return err
-		}
-	}
 
 	// If PVC' still exists, delete it
 	if pvcPrime != nil {
